Extract sidecar save-path store
