+++ docs/dev/plans/refactor-4-inventory.md
@@ -44,7 +44,7 @@ locked KSP process/log condition above; the build itself succeeds.
 | `Source/Parsek/ParsekFlight.cs` | 14,503 | Pass1-Done; post-switch auto-record trigger helpers extracted, finalization deferred |
 | `Source/Parsek/GhostVisualBuilder.cs` | 7,193 | Pass1-Deferred; visual-builder split needs owner plan and runtime validation |
 | `Source/Parsek/GameActions/LedgerOrchestrator.cs` | 6,976 | Pass1-Done; earnings-window, vessel-cost, and recalculation helpers extracted |
-| `Source/Parsek/RecordingStore.cs` | 6,834 | Pass2-Done for `SidecarFileCommitBatch`; Pass1-Done for optimization and rewind helpers; remaining sidecar orchestration deferred |
+| `Source/Parsek/RecordingStore.cs` | 6,365 | Pass2-Done for `SidecarFileCommitBatch` and save-path `RecordingSidecarStore`; load-path orchestration deferred |
 | `Source/Parsek/FlightRecorder.cs` | 6,689 | Pass1-Done; visual coverage logging helpers extracted |
 | `Source/Parsek/GhostPlaybackLogic.cs` | 5,343 | Pass1-Done; ghost info population and part-event helpers extracted |
 | `Source/Parsek/UI/RecordingsTableUI.cs` | 4,868 | Pass1-Deferred; high-coupling IMGUI row/tree split deferred |
@@ -71,7 +71,7 @@ These deltas compare files that also existed in
 |------|------------------|---------------|-------|
 | `LedgerOrchestrator.cs` | 900 | 6,976 | +6,076 |
 | `ParsekFlight.cs` | 8,765 | 14,503 | +5,738 |
-| `RecordingStore.cs` | 2,958 | 6,834 | +3,876 |
+| `RecordingStore.cs` | 2,958 | 6,365 | +3,407 |
 | `GhostPlaybackLogic.cs` | 2,589 | 5,343 | +2,754 |
 | `VesselSpawner.cs` | 1,473 | 4,166 | +2,693 |
 | `GhostPlaybackEngine.cs` | 1,770 | 4,312 | +2,542 |
@@ -141,7 +141,7 @@ central enough that parallel edits would make review and rollback worse.
 |------|--------|
 | `ParsekFlight.cs` | 14.5k-line scene controller, +5.7k since refactor-3 |
 | `LedgerOrchestrator.cs` | 7.0k-line GameActions hub, +6.1k since refactor-3 |
-| `RecordingStore.cs` | 6.8k-line storage surface, +3.9k since refactor-3; Pass 2 sidecar commit batch extracted |
+| `RecordingStore.cs` | 6.4k-line storage surface, +3.4k since refactor-3; Pass 2 sidecar commit batch and save path extracted |
 | `FlightRecorder.cs` | 6.7k-line sampling/event surface |
 | `GhostPlaybackLogic.cs` | 5.3k-line playback/visual logic helper |
 | `UI/RecordingsTableUI.cs` | 4.9k-line extracted UI surface with prior coupling risk |
@@ -258,7 +258,7 @@ loops can fool a mechanical scan.
 | `ParsekScenario.cs` | `LoadRecordingMetadata` | 3660 | 244 |
 | `RecordingStore.cs` | `RunOptimizationPass` | 1912 | 216 |
 | `RecordingStore.cs` | `InitiateRewind` | 3464 | 141 |
-| `RecordingStore.cs` | `ReconcileReadableSidecarMirrors` | 6413 | 128 |
+| `RecordingSidecarStore.cs` | `ReconcileReadableSidecarMirrors` | 287 | 119 |
 | `UI/RecordingsTableUI.cs` | `DrawRecordingsTableHeader` | 835 | 136 |
 | `UI/RecordingsTableUI.cs` | `DrawRecordingsWindow` | 1045 | 232 |
 | `UI/RecordingsTableUI.cs` | `DrawRecordingRow` | 1281 | 382 |
@@ -455,19 +455,31 @@ Pass 2 first slice completed:
 
 - Extracted `SidecarFileCommitBatch` into `Source/Parsek/SidecarFileCommitBatch.cs`
   for staged sidecar write/delete commits, rollback, and artifact cleanup.
-  `RecordingStore` still owns save/load orchestration, sidecar epoch ownership
-  and mutation order, `FilesDirty`, readable mirrors, snapshot policy, and codec
-  dispatch.
+  At this checkpoint, `RecordingStore` still owned save/load orchestration,
+  sidecar epoch ownership and mutation order, `FilesDirty`, readable mirrors,
+  snapshot policy, and codec dispatch.
+
+Pass 2 second slice completed:
+
+- Extracted save-path `RecordingSidecarStore` into
+  `Source/Parsek/RecordingSidecarStore.cs` for save-side path resolution,
+  sidecar epoch bump/rollback, staged authoritative sidecar writes, readable
+  mirror reconciliation, and `FilesDirty` clearing. `RecordingStore` keeps the
+  existing save wrappers and still owns load-path hydration, sidecar epoch
+  validation, loop migration/repair, terminal-orbit and endpoint backfills,
+  snapshot fallback/failure policy, grouping, optimization, deletion, and rewind
+  entry points.
 
 Validation:
 
 - `dotnet build Source/Parsek/Parsek.csproj`
 - `dotnet test Source/Parsek.Tests/Parsek.Tests.csproj --filter "FullyQualifiedName~RecordingOptimizerTests|FullyQualifiedName~RecordingStoreTests|FullyQualifiedName~LegacyTreeMigrationTests|FullyQualifiedName~RewindLoggingTests"`
 - `dotnet test Source/Parsek.Tests/Parsek.Tests.csproj --filter FullyQualifiedName!~InjectAllRecordings`
+- Pass 2 save-path slice: `dotnet test Source/Parsek.Tests/Parsek.Tests.csproj --filter "FullyQualifiedName~RecordingStorageRoundTripTests|FullyQualifiedName~SnapshotSidecarCodecTests|FullyQualifiedName~TrajectorySidecarBinaryTests|FullyQualifiedName~Bug270SidecarEpochTests|FullyQualifiedName~FormatVersionTests|FullyQualifiedName~TrackSectionSerializationTests|FullyQualifiedName~LoopIntervalLoadNormalizationTests|FullyQualifiedName~QuickloadResumeTests"` passed 235 tests; `dotnet test Source/Parsek.Tests/Parsek.Tests.csproj --filter FullyQualifiedName!~InjectAllRecordings` passed 8,647 tests; `dotnet test Source/Parsek.Tests/Parsek.Tests.csproj --filter FullyQualifiedName~InjectAllRecordings` passed 3 tests.
 
-Remaining sidecar orchestration and codec work follows the Pass 2 owner plan
-and remains gated on preserving file ordering, exception handling, sidecar epoch
-ordering, and `FilesDirty` mutation order exactly.
+Remaining load-path sidecar orchestration and codec work follows the Pass 2
+owner plan and remains gated on preserving file ordering, exception handling,
+sidecar epoch ordering, and `FilesDirty` mutation order exactly.
 
 Pass 2 storage/sidecar owner proposal:
 `docs/dev/plans/refactor-4-pass2-storage-sidecars.md`. Rewind service
@@ -910,7 +922,7 @@ semantic, architectural, runtime-visual, math-sensitive, or UI-order-sensitive.
 | `ParsekFlight.cs` | Done for post-switch auto-record; finalization split deferred to Pass 2. |
 | `FlightRecorder.cs` | Done for visual coverage logging; remaining part-event poller work deferred. |
 | `GhostPlaybackLogic.cs` | Done for dictionary population and part events; remaining spawn policy cleanup deferred. |
-| `RecordingStore.cs` | Pass 2 first slice done for `SidecarFileCommitBatch`; save/load orchestration, codecs, grouping, optimization, deletion, and rewind wrappers remain with `RecordingStore` until separately approved. |
+| `RecordingStore.cs` | Pass 2 first and second slices done for `SidecarFileCommitBatch` and save-path `RecordingSidecarStore`; load orchestration, codecs, grouping, optimization, deletion, and rewind wrappers remain with `RecordingStore` until separately approved. |
 | `GameStateRecorder.cs` | Deferred; resource/milestone/facility handler families need owner map. |
 | `GameActions/KspStatePatcher.cs` | Deferred; patch-order/reflection/UI mutation paths need state-family patcher proposal. |
 | `BallisticExtrapolator.cs` | Deferred; math and iteration-order sensitive. |
@@ -956,9 +968,9 @@ raw scan with a manual map for the high-risk owners:
    deduplication. Initial result: `refactor-4-pass2-storage-sidecars.md`
    recommends separate sidecar commit, sidecar orchestration, trajectory text
    codec, manifest codec, and tree-record codec owners. The sidecar commit
-   batch slice is complete; schema redesign remains out of scope, and binary/text
-   format unification is deferred until the orchestration and manifest codec
-   owners land.
+   batch and save-path orchestration slices are complete; schema redesign
+   remains out of scope, and binary/text format unification is deferred until
+   the load-path orchestration and manifest codec owners land.
 3. Build a static mutable state map for `GameStateRecorder`,
    `LedgerOrchestrator`, `RecordingStore`, `ParsekScenario`,
    `WatchModeController`, and `GhostPlaybackEngine`.
